Products Page table done
finished up the table. need to set it up as a loop. i don’t know if the
cart is set up fully. just need to loop thru that to show everything.

also need to add the form at the bottom for the user to fill out. not
sure how that information is being sent to the database, but will need
to set that up too.